+++ packages/plugin-packages/mashroom-security-provider-openid-connect/test/create-user.test.ts
@@ -1,11 +1,11 @@
 
 import createUser from '../src/create-user';
-import type {IdTokenClaims, UserinfoResponse} from 'openid-client';
+import type {IDToken, UserInfoResponse} from 'openid-client';
 
 describe('create-user', () => {
 
     it('creates a user from IDToken claims that only contains sub', () => {
-        const claims: IdTokenClaims = {
+        const claims: IDToken = {
             aud: 'x',
             exp: 0,
             iat: 0,
@@ -27,7 +27,7 @@ describe('create-user', () => {
     });
 
     it('creates a user from IDToken claims', () => {
-        const claims: IdTokenClaims = {
+        const claims: IDToken = {
             aud: 'x',
             exp: 0,
             iat: 0,
@@ -52,7 +52,7 @@ describe('create-user', () => {
     });
 
     it('gets the roles from the IdToken claims', () => {
-        const claims: IdTokenClaims = {
+        const claims: IDToken = {
             aud: 'x',
             exp: 0,
             iat: 0,
@@ -78,14 +78,14 @@ describe('create-user', () => {
     });
 
     it('creates a user from user info', () => {
-        const claims: IdTokenClaims = {
+        const claims: IDToken = {
             aud: 'x',
             exp: 0,
             iat: 0,
             iss: 'x',
             sub: 'admin',
         };
-        const userInfo: UserinfoResponse = {
+        const userInfo: UserInfoResponse = {
             sub: 'admin',
             preferred_username: 'admin',
             name: 'Admin User',
@@ -108,7 +108,7 @@ describe('create-user', () => {
     });
 
     it('copies extra data from the id token', () => {
-        const claims: IdTokenClaims = {
+        const claims: IDToken = {
             aud: 'x',
             exp: 0,
             iat: 0,
@@ -141,14 +141,14 @@ describe('create-user', () => {
     });
 
     it('copies extra data from the user info', () => {
-        const claims: IdTokenClaims = {
+        const claims: IDToken = {
             aud: 'x',
             exp: 0,
             iat: 0,
             iss: 'x',
             sub: 'admin',
         };
-        const userInfo: UserinfoResponse = {
+        const userInfo: UserInfoResponse = {
             sub: 'admin',
             preferred_username: 'admin',
             name: 'Admin User',