+++ packages/plugin-packages/mashroom-security-provider-openid-connect/test/security-provider/MashroomOpenIDConnectSecurityProvider.test.ts
@@ -6,7 +6,7 @@ describe('MashroomOpenIDConnectSecurityProvider', () => {
 
     it('returns the user if the token is not expired', () => {
 
-        const provider = new MashroomOpenIDConnectSecurityProvider('');
+        const provider = new MashroomOpenIDConnectSecurityProvider({} as any, '');
 
         const req: any = {
             session: {
@@ -47,7 +47,7 @@ describe('MashroomOpenIDConnectSecurityProvider', () => {
 
     it('does not return the user if the token is expired', () => {
 
-        const provider = new MashroomOpenIDConnectSecurityProvider('');
+        const provider = new MashroomOpenIDConnectSecurityProvider({} as any, '');
 
         const req: any = {
             session: {